+++ src/js/config/appConfig.js
@@ -7,8 +7,14 @@ function appConfig($routeProvider) {
     .when("/", {
       templateUrl: "/dist/partials/index.template.html"
     })
-    .when("/post/:year/:month/:date/:title", {
-      templateUrl: (params) => `/dist/partials/${params.year}/${params.month}/${params.date}/${params.title}.template.html`,
+    .when("/posts", {
+      templateUrl: "/dist/partials/postlist.template.html",
+      controller: "PostListController",
+      controllerAs: "vm",
+      bindToController: true
+    })
+    .when("/posts/:year/:month/:date/:title", {
+      templateUrl: (params) => `/dist/partials/posts/${params.year}/${params.month}/${params.date}/${params.title}.template.html`,
       controller: "PostController",
       controllerAs: "vm",
       bindToController: true