+++ src/js/routes/postlist/postlist.module.js
@@ -1,0 +1,4 @@
+import PostListController from "./postlist.controller";
+
+angular.module("ghpages.postlist", [])
+  .controller("PostListController", PostListController)